player class added and integrated
